@@ -16,7 +16,7 @@
 
 package org.springframework.samples.petclinic.rest;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
