@@ -4,8 +4,8 @@ import co.elastic.apm.api.CaptureSpan;
 import co.elastic.apm.api.ElasticApm;
 import co.elastic.apm.api.Transaction;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
